# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Raster-to-Vector API
  description: |
    Enterprise-grade raster-to-vector conversion API that transforms images into scalable vector formats.
    
    ## Overview
    Convert PNG, JPEG, BMP, and TIFF images to high-quality SVG vectors using multiple vectorization engines including VTracer, OpenCV, Potrace, Inkscape, AutoTrace, Primitive, and SVGCleaner.
    
    ## Features
    - Multiple vectorization algorithms for different use cases
    - Asynchronous job processing with status tracking
    - Batch conversion support
    - Comprehensive parameter customization
    - File management with metadata extraction
    - Production-ready error handling
    
    ## Base URL
    ```
    https://api.yourservice.com/v1
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    url: https://yourservice.com/support
    email: support@yourservice.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://yourservice.com/terms
servers:
  - url: https://api.yourservice.com/v1
    description: Production server
  - url: http://localhost:3002/api
    description: Development server
paths:
...
  /files/{fileId}:
    get:
      summary: Get file information
      description: Retrieve metadata and information about an uploaded file
      operationId: getFileInfo
      tags:
        - Files
      parameters:
        - name: fileId
          in: path
          required: true
          description: Unique identifier of the file
          schema:
            type: string
      responses:
        '200':
          description: File information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileInfoResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Delete file
      description: Delete an uploaded file and its metadata
      operationId: deleteFile
      tags:
        - Files
      parameters:
        - name: fileId
          in: path
          required: true
          description: Unique identifier of the file
          schema:
            type: string
      responses:
        '200':
          description: File deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /convert/{jobId}:
    get:
      summary: Get job details
      description: Retrieve complete information about a conversion job
      operationId: getJobDetails
      tags:
        - Conversions
      parameters:
        - name: jobId
          in: path
          required: true
          description: Unique identifier of the conversion job
          schema:
            type: string
      responses:
        '200':
          description: Job details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetailsResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Cancel conversion job
      description: Cancel a running or queued conversion job
      operationId: cancelJob
      tags:
        - Conversions
      parameters:
        - name: jobId
          in: path
          required: true
          description: Unique identifier of the conversion job
          schema:
            type: string
      responses:
        '200':
          description: Job cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    FileInfoResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              example: "file_abc123"
            originalName:
              type: string
              example: "logo.png"
            filename:
              type: string
              example: "abc123_logo.png"
            mimetype:
              type: string
              example: "image/png"
            size:
              type: integer
              example: 52428
            path:
              type: string
              example: "/uploads/abc123_logo.png"
            metadata:
              $ref: '#/components/schemas/ImageMetadata'
            createdAt:
              type: string
              format: date-time
              example: "2025-08-12T14:30:00Z"
    ImageMetadata:
      type: object
      properties:
        width:
          type: integer
          description: Image width in pixels
          example: 1024
        height:
          type: integer
          description: Image height in pixels
          example: 768
        channels:
          type: integer
          description: Number of color channels
          example: 3
        colorSpace:
          type: string
          description: Color space of the image
          example: "srgb"
        hasAlpha:
          type: boolean
          description: Whether image has transparency
          example: false
        density:
          type: number
          description: Pixel density (DPI)
          example: 72
    
    # Method schemas
...
    JobDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Job'
...
    Job:
      type: object
      properties:
        id:
          type: string
          description: Unique job identifier
          example: "job_xyz789"
        fileId:
          type: string
          description: ID of the input file
          example: "file_abc123"
        method:
          type: string
          description: Vectorization method used
          example: "vtracer"
        parameters:
          type: object
          description: Parameters used for conversion
          additionalProperties: true
        status:
          type: string
          enum: [queued, processing, completed, failed, cancelled]
          example: "completed"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          example: 100
        createdAt:
          type: string
          format: date-time
          example: "2025-08-12T14:30:00Z"
        startedAt:
          type: string
          format: date-time
          example: "2025-08-12T14:31:00Z"
        completedAt:
          type: string
          format: date-time
          example: "2025-08-12T14:32:00Z"
        estimatedTime:
          type: integer
          description: Estimated completion time in milliseconds
          example: 5000
        resultPath:
          type: string
          description: Path to result file (when completed)
          example: "/outputs/job_xyz789.svg"
        qualityMetrics:
          $ref: '#/components/schemas/QualityMetrics'
        error:
          type: string
          description: Error message if job failed
    QualityMetrics:
      type: object
      properties:
        pathCount:
          type: integer
          description: Number of SVG paths generated
          example: 45
        pointCount:
          type: integer
          description: Number of control points
          example: 180
        fileSize:
          type: integer
          description: Output file size in bytes
          example: 8192
        processingTime:
          type: integer
          description: Processing time in milliseconds
          example: 2500
        accuracy:
          type: number
          minimum: 0
          maximum: 1
          description: Estimated accuracy score
          example: 0.92
        smoothness:
          type: number
          minimum: 0
          maximum: 1
          description: Path smoothness score
          example: 0.88
    
    # Generic response schemas
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        requestId:
          type: string
          description: Unique request identifier for tracing
          example: "550e8400-e29b-41d4-a716-446655440000"
    ErrorResponse:
      type: object
      description: "RFC 7807 Problem Details for HTTP APIs"
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
          example: "https://api.yourservice.com/problems/validation-error"
        title:
          type: string
          description: A short, human-readable summary of the problem type
          example: "Validation Error"
        status:
          type: integer
          description: The HTTP status code for this occurrence of the problem
          example: 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
          example: "The 'fileId' parameter is required but was not provided"
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
          example: "/api/convert"
        requestId:
          type: string
          description: Unique request identifier for tracing
          example: "550e8400-e29b-41d4-a716-446655440000"
      required:
        - type
        - title
        - status
...